+++ Package@swift-5.swift
@@ -6,8 +6,8 @@ let package = Package(
     name: "QuickConstraint",
     platforms: [
         .macOS(.v10_14),
-        .iOS(.v9),
-        .tvOS(.v9)
+        .iOS(.v10),
+        .tvOS(.v10)
     ],
     products: [
         .library(name: "QuickConstraint",